Moved adapters to dedicated namespace.
