+++ docs/features/base64-upload-adapter.md
@@ -6,7 +6,7 @@ order: 40
 
 # Base64 upload adapter
 
-The {@link module:upload/base64uploadadapter~Base64UploadAdapter Base64 image upload adapter} plugin converts images inserted into the editor into [Base64 strings](https://en.wikipedia.org/wiki/Base64) in the {@link builds/guides/integration/saving-data editor output}.
+The {@link module:upload/adapters/base64uploadadapter~Base64UploadAdapter Base64 image upload adapter} plugin converts images inserted into the editor into [Base64 strings](https://en.wikipedia.org/wiki/Base64) in the {@link builds/guides/integration/saving-data editor output}.
 
 This kind of image upload does not require server processing – images are stored with the rest of the text and displayed by the web browser without additional requests. On the downside, this approach can bloat your database with very long data strings which, in theory, could have a negative impact on the performance.
 
@@ -32,7 +32,7 @@ To add this feature to your editor, install the [`@ckeditor/ckeditor5-upload`](h
 npm install --save @ckeditor/ckeditor5-upload
 ```
 
-Then add the {@link module:upload/base64uploadadapter~Base64UploadAdapter `Base64UploadAdapter`} to your plugin list:
+Then add the {@link module:upload/adapters/base64uploadadapter~Base64UploadAdapter `Base64UploadAdapter`} to your plugin list:
 
 ```js
 import Base64UploadAdapter from '@ckeditor/ckeditor5-upload/src/base64uploadadapter';
