+++ tests/adapters/simpleuploadadapter.js
@@ -6,8 +6,8 @@
 /* globals document */
 
 import ClassicTestEditor from '@ckeditor/ckeditor5-core/tests/_utils/classictesteditor';
-import SimpleUploadAdapter from '../src/simpleuploadadapter';
-import FileRepository from '../src/filerepository';
+import SimpleUploadAdapter from '../../src/adapters/simpleuploadadapter';
+import FileRepository from '../../src/filerepository';
 import log from '@ckeditor/ckeditor5-utils/src/log';
 import { createNativeFileMock } from '@ckeditor/ckeditor5-upload/tests/_utils/mocks';
 import testUtils from '@ckeditor/ckeditor5-core/tests/_utils/utils';
